Tuanndph31135

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\Controller;
 use App\Models\Category;
-use App\Http\Requests\StoreCategoryRequest;
-// use App\Http\Requests\UpdateCategoryRequest;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+// use App\Http\Requests\UpdateCategoryRequest;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\StoreCategoryRequest;
 
 class CategoryController extends Controller
 {
@@ -18,9 +19,13 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $data = Category::query()->with(['parent', 'children'])->latest('id')->paginate(5);
-
-        return view(self::PATH_VIEW . __FUNCTION__, compact('data'));
+        $user = Auth::user();
+        $data = Category::query()->with(['parent', 'children'])->latest('id')->paginate(10);
+        if ($user->role_id === 1) {
+            return view(self::PATH_VIEW . __FUNCTION__, compact('data'));
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 
     /**
@@ -28,21 +33,29 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $user = Auth::user();
         $parentCategories = Category::query()->with('children')->whereNull('parent_id')->get();
 
         // dd($parentCategories);
-
-        return view(self::PATH_VIEW . __FUNCTION__, compact('parentCategories'));
+        if ($user->role_id === 1) {
+            return view(self::PATH_VIEW . __FUNCTION__, compact('parentCategories'));
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(StoreCategoryRequest $request)
     {
+        $user = Auth::user();
         Category::query()->create($request->all());
-
-        return redirect()->route('admin.categories.index');
+        if ($user->role_id === 1) {
+            return redirect()->route('admin.categories.index')->with('success', 'Category created successfully!');;
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 
     /**
@@ -50,9 +63,13 @@ class CategoryController extends Controller
      */
     public function show(string $id)
     {
+        $user = Auth::user();
         $model = Category::query()->findOrFail($id);
-
-        return view(self::PATH_VIEW . __FUNCTION__, compact('model'));
+        if ($user->role_id === 1) {
+            return view(self::PATH_VIEW . __FUNCTION__, compact('model'));
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 
     /**
@@ -60,11 +77,15 @@ class CategoryController extends Controller
      */
     public function edit(string $id)
     {
+        $user = Auth::user();
         $model = Category::query()->findOrFail($id);
 
         $parentCategories = Category::query()->with('children')->whereNull('parent_id')->get();
-
-        return view(self::PATH_VIEW . __FUNCTION__, compact('model', 'parentCategories'));
+        if ($user->role_id === 1) {
+            return view(self::PATH_VIEW . __FUNCTION__, compact('model', 'parentCategories'));
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 
     /**
@@ -72,11 +93,15 @@ class CategoryController extends Controller
      */
     public function update(StoreCategoryRequest $request, string $id)
     {
+        $user = Auth::user();
         $model = Category::query()->findOrFail($id);
 
         $model->update($request->all());
-
-        return redirect()->route('admin.categories.index');
+        if ($user->role_id === 1) {
+            return redirect()->route('admin.categories.index')->with('success', 'Category updated successfully!');
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 
     /**
@@ -84,10 +109,14 @@ class CategoryController extends Controller
      */
     public function destroy(string $id)
     {
+        $user = Auth::user();
         $model = Category::query()->findOrFail($id);
+        if ($user->role_id === 1) {
+            $model->delete();
 
-        $model->delete();
-
-        return back();
+            return back()->with('success', 'Category deleted successfully!');
+        } else {
+            return back()->with('error', 'Access denied!');
+        };
     }
 }
